Minor float/float32/double/float64 cleanup (#695)
* Remove noise casts
* Use FLT_EPSILON and DBL_EPSILON instead of misc epsilon guesses
* Add float suffixes and remove casts

+++ hdf/test/man.c
@@ -117,7 +117,7 @@ genimage(int height, int width, float32 *data, uint8 *image)
 
     /* store one value per row, increasing by one for each row */
     pdata      = data;
-    multiplier = (float32)255.0 / (max - min);
+    multiplier = 255.0F / (max - min);
     for (i = 0; i < limit; i++)
         *image++ = (uint8)(((*pdata++) - min) * multiplier);
 } /* geniamge() */
